refactor(update): split UpdateApplicator into IUpdateApplicator + 3 platforms
The 320-line `object UpdateApplicator` had three platform-specific update
flows + shared helpers all packed together. Same smell that triggered
the original Config-split for AppConfig: unmockable singleton with five
unrelated responsibilities.

Structure now:
  - `IUpdateApplicator` interface in client-core/.../api/interfaces/
    (so UpdateDialog in client-ui can inject the interface without
    pulling in concrete platform impls; layering stays one-way).
  - WindowsUpdateApplicator / MacUpdateApplicator / LinuxUpdateApplicator
    in client-launcher/.../update/ — each owns its own resolveExecutable,
    shutdown-hook body, and rollback logic.
  - NoOpUpdateApplicator throws UnsupportedOperationException for
    unrecognised hosts; the factory branch stays exhaustive.
  - UpdateApplicators.forCurrentPlatform() picks by OS at startup.
    Singleton in Koin so each impl's shutdown hook registers exactly once.

OS detection extracted from the bottom of the old UpdateApplicator.kt
into platform/OS.kt — sits next to PlatformPaths and SingleInstance,
which is where every "what is this machine" concern now lives.

UpdateDialog switched from `UpdateApplicator.scheduleUpdate(...)` static
to `koinInject<IUpdateApplicator>()` — same call site, injected concrete.

One test (UpdateApplicatorsTest.factory-returns-host-matching-impl) for
the factory branch logic; the platform-specific scheduleUpdate methods
remain hard to cover without spawning real processes — that's the
existing limitation, not a regression from this split.
